FIxed a bug where uncategorised transactions had to be coded even though they were outside the current reconciliation period. Now only transactions within the period need to be coded.
