+++ BudgetAnalyser.Engine/Services/LedgerService.cs
@@ -1,6 +1,8 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
+using System.Globalization;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -167,7 +169,7 @@ namespace BudgetAnalyser.Engine.Services
         }
 
         public LedgerEntryLine MonthEndReconciliation(
-            DateTime reconciliationStartDate,
+            DateTime reconciliationDate,
             IEnumerable<BankBalance> balances,
             IBudgetCurrencyContext budgetContext,
             StatementModel statement,
@@ -196,7 +198,32 @@ namespace BudgetAnalyser.Engine.Services
             ReconciliationToDoList.Clear();
             Stopwatch stopWatch = Stopwatch.StartNew();
             this.logger.LogInfo(l => l.Format("Starting Ledger Book reconciliation {0}", DateTime.Now));
-            LedgerEntryLine recon = LedgerBook.Reconcile(reconciliationStartDate, balances, budgetContext.Model, ReconciliationToDoList, statement, ignoreWarnings);
+
+            try
+            {
+                PreReconciliationValidation(reconciliationDate, statement);
+            }
+            catch (ValidationWarningException)
+            {
+                if (!ignoreWarnings)
+                {
+                    throw;
+                }
+            }
+
+            if (ReconciliationToDoList == null)
+            {
+                ReconciliationToDoList = new ToDoCollection();
+            }
+
+            decimal consistencyCheck1 = LedgerBook.Reconciliations.Sum(e => e.CalculatedSurplus);
+            LedgerEntryLine recon = LedgerBook.Reconcile(reconciliationDate, balances, budgetContext.Model, ReconciliationToDoList, statement);
+            decimal consistencyCheck2 = LedgerBook.Reconciliations.Sum(e => e.CalculatedSurplus);
+            if (consistencyCheck1 != consistencyCheck2)
+            {
+                throw new CorruptedLedgerBookException("Code Error: The previous dated entries have changed, this is not allowed. Data is corrupt.");
+            }
+
             foreach (ToDoTask task in ReconciliationToDoList)
             {
                 this.logger.LogInfo(l => l.Format("TASK: {0} SystemGenerated:{1}", task.Description, task.SystemGenerated));
@@ -331,6 +358,121 @@ namespace BudgetAnalyser.Engine.Services
         public IEnumerable<Account.Account> ValidLedgerAccounts()
         {
             return this.accountTypeRepository.ListCurrentlyUsedAccountTypes();
+        }
+
+        private void PreReconciliationValidation(DateTime reconciliationDate, StatementModel statement)
+        {
+            var messages = new StringBuilder();
+            if (!LedgerBook.Validate(messages))
+            {
+                throw new InvalidOperationException("Ledger book is currently in an invalid state. Cannot add new entries.\n" + messages);
+            }
+
+            if (statement == null)
+            {
+                return;
+            }
+
+            var startDate = ReconciliationBuilder.CalculateDateForReconcile(LedgerBook, reconciliationDate);
+
+            ValidateDates(startDate, reconciliationDate, statement);
+
+            ValidateAgainstUncategorisedTransactions(startDate, reconciliationDate, statement);
+
+            ValidateAgainstOrphanedAutoMatchingTransactions(statement);
+        }
+
+        private void ValidateAgainstOrphanedAutoMatchingTransactions(StatementModel statement)
+        {
+            LedgerEntryLine lastLine = LedgerBook.Reconciliations.FirstOrDefault();
+            if (lastLine == null)
+            {
+                return;
+            }
+
+            List<LedgerTransaction> unmatchedTxns = lastLine.Entries
+                .SelectMany(e => e.Transactions)
+                .Where(t => !string.IsNullOrWhiteSpace(t.AutoMatchingReference) && !t.AutoMatchingReference.StartsWith(ReconciliationBuilder.MatchedPrefix, StringComparison.Ordinal))
+                .ToList();
+
+            if (unmatchedTxns.None())
+            {
+                return;
+            }
+
+            List<Transaction> statementSubSet = statement.AllTransactions.Where(t => t.Date >= lastLine.Date).ToList();
+            foreach (LedgerTransaction ledgerTransaction in unmatchedTxns)
+            {
+                IEnumerable<Transaction> statementTxns = ReconciliationBuilder.TransactionsToAutoMatch(statementSubSet, ledgerTransaction.AutoMatchingReference);
+                if (statementTxns.None())
+                {
+                    this.logger.LogWarning(
+                        l =>
+                            l.Format(
+                                "There appears to be some transactions from last month that should be auto-matched to a statement transactions, but no matching statement transactions were found. {0}",
+                                ledgerTransaction));
+                    throw new ValidationWarningException(
+                        string.Format(
+                            CultureInfo.CurrentCulture,
+                            "There appears to be some transactions from last month that should be auto-matched to a statement transactions, but no matching statement transactions were found.\nHave you forgotten to do a transfer?\nTransaction ID:{0} Ref:{1} Amount:{2:C}",
+                            ledgerTransaction.Id,
+                            ledgerTransaction.AutoMatchingReference,
+                            ledgerTransaction.Amount));
+                }
+            }
+        }
+
+        private void ValidateAgainstUncategorisedTransactions(DateTime startDate, DateTime reconciliationDate, StatementModel statement)
+        {
+            if (statement.AllTransactions
+                .Where(t => t.Date >= startDate && t.Date < reconciliationDate)
+                .Any(t => t.BudgetBucket == null || (t.BudgetBucket != null && string.IsNullOrWhiteSpace(t.BudgetBucket.Code))))
+            {
+                IEnumerable<Transaction> uncategorised = statement.AllTransactions.Where(t => t.BudgetBucket == null || (t.BudgetBucket != null && string.IsNullOrWhiteSpace(t.BudgetBucket.Code)));
+                var count = 0;
+                this.logger.LogWarning(_ => "LedgerBook.PreReconciliationValidation: There appears to be transactions in the statement that are not categorised into a budget bucket.");
+                foreach (Transaction transaction in uncategorised)
+                {
+                    count++;
+                    Transaction transactionCopy = transaction;
+                    this.logger.LogWarning(_ => "LedgerBook.PreReconciliationValidation: Transaction: " + transactionCopy.Id + transactionCopy.BudgetBucket);
+                    if (count > 5)
+                    {
+                        this.logger.LogWarning(_ => "LedgerBook.PreReconciliationValidation: There are more than 5 transactions.");
+                    }
+                }
+
+                throw new ValidationWarningException("There appears to be transactions in the statement that are not categorised into a budget bucket.");
+            }
+        }
+
+        [SuppressMessage("ReSharper", "UnusedParameter.Local")]
+        private void ValidateDates(DateTime startDate, DateTime reconciliationDate, StatementModel statement)
+        {
+            LedgerEntryLine recentEntry = LedgerBook.Reconciliations.FirstOrDefault();
+            if (recentEntry != null)
+            {
+                if (reconciliationDate <= recentEntry.Date)
+                {
+                    throw new InvalidOperationException("The start Date entered is before the previous ledger entry.");
+                }
+
+                if (recentEntry.Date.AddDays(7 * 4) > reconciliationDate)
+                {
+                    throw new InvalidOperationException("The start Date entered is not at least 4 weeks after the previous reconciliation. ");
+                }
+
+                if (recentEntry.Date.Day != reconciliationDate.Day)
+                {
+                    throw new ValidationWarningException(
+                        "The reconciliation Date chosen, {0}, isn't the same day of the month as the previous entry {1}. Not required, but ideally reconciliations should be evenly spaced.");
+                }
+            }
+
+            if (!statement.AllTransactions.Any(t => t.Date >= startDate))
+            {
+                throw new ValidationWarningException("There doesn't appear to be any transactions in the statement for the month up to " + reconciliationDate.ToShortDateString());
+            }
         }
     }
 }
